+++ modules/shared_data/src/test/scala/com/my/dor_metagraph/shared_data/combiners/StakingCombinerTest.scala
@@ -80,7 +80,7 @@ object StakingCombinerTest extends MutableIOSuite {
       tokenA,
       tokenB,
       owner,
-      (tokenA.amount.value * tokenB.amount.value).toDouble,
+      BigInt(tokenA.amount.value) * BigInt(tokenB.amount.value),
       PoolShares(totalShares, shares)
     )
     (
